+++ src/__tests__/opportunities.int.test.ts
@@ -11,12 +11,8 @@ import {
 	loadSystemUser,
 } from '../database';
 import { createTestFunder, createTestOpportunity } from '../test/factories';
-import {
-	getAuthContext,
-	loadTestUser,
-	NO_LIMIT,
-	NO_OFFSET,
-} from '../test/utils';
+import { NO_LIMIT, NO_OFFSET } from '../constants';
+import { getAuthContext, loadTestUser } from '../test/utils';
 import {
 	expectArray,
 	expectArrayContaining,