+++ src/__tests__/proposals.int.test.ts
@@ -24,12 +24,8 @@ import {
 	createTestProposal,
 	createTestUser,
 } from '../test/factories';
-import {
-	getAuthContext,
-	loadTestUser,
-	NO_LIMIT,
-	NO_OFFSET,
-} from '../test/utils';
+import { NO_LIMIT, NO_OFFSET } from '../constants';
+import { getAuthContext, loadTestUser } from '../test/utils';
 import {
 	expectArray,
 	expectArrayContaining,