+++ src/__tests__/sources.int.test.ts
@@ -27,12 +27,11 @@ import {
 	createTestProposal,
 	createTestSource,
 } from '../test/factories';
+import { NO_LIMIT, NO_OFFSET } from '../constants';
 import {
 	getAuthContext,
 	getTestAuthContext,
 	loadTestUser,
-	NO_LIMIT,
-	NO_OFFSET,
 } from '../test/utils';
 import {
 	mockJwt as authHeader,