+++ src/__tests__/terminologySets.int.test.ts
@@ -12,12 +12,8 @@ import {
 	loadTerminologySet,
 } from '../database';
 import { createTestFunder, createTestTerminologySet } from '../test/factories';
-import {
-	getAuthContext,
-	loadTestUser,
-	NO_LIMIT,
-	NO_OFFSET,
-} from '../test/utils';
+import { NO_LIMIT, NO_OFFSET } from '../constants';
+import { getAuthContext, loadTestUser } from '../test/utils';
 import {
 	expectArray,
 	expectArrayContaining,