+++ src/constants/index.ts
@@ -1,5 +1,6 @@
 export * from './exitCodes';
 export * from './httpStatus';
 export * from './numbers';
+export * from './pagination';
 export * from './testDatabasePrefix';
 export * from './time';